+++ cni/pkg/nodeagent/server.go
@@ -23,6 +23,7 @@ import (
 	"path/filepath"
 	"sync/atomic"
 
+	"golang.org/x/sys/unix"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
@@ -33,6 +34,7 @@ import (
 	"istio.io/istio/cni/pkg/scopes"
 	"istio.io/istio/cni/pkg/util"
 	"istio.io/istio/pkg/kube"
+	"istio.io/istio/pkg/util/sets"
 )
 
 var log = scopes.CNIAgent
@@ -44,7 +46,6 @@ type MeshDataplane interface {
 
 	//	IsPodInMesh(ctx context.Context, pod *metav1.ObjectMeta, netNs string) (bool, error)
 	AddPodToMesh(ctx context.Context, pod *corev1.Pod, podIPs []netip.Addr, netNs string) error
-	AddPodToHostNSIpset(pod *corev1.Pod, podIPs []netip.Addr) ([]netip.Addr, error)
 	RemovePodFromMesh(ctx context.Context, pod *corev1.Pod, isDelete bool) error
 
 	Stop()
@@ -109,6 +110,8 @@ func NewServer(ctx context.Context, ready *atomic.Value, pluginSocket string, ar
 		dataplane: &meshDataplane{
 			kubeClient: client.Kube(),
 			netServer:  netServer,
+			hostIptables:       hostIptables,
+			hostsideProbeIPSet: set,
 		},
 	}
 	s.NotReady()
@@ -182,8 +185,6 @@ func (s *Server) Start() {
 }
 
 func (s *Server) Stop() {
-	log.Info("CNI ambient server terminating, cleaning up node net rules")
-
 	s.cniServerStopFunc()
 	s.dataplane.Stop()
 }
@@ -191,6 +192,8 @@ func (s *Server) Stop() {
 type meshDataplane struct {
 	kubeClient kubernetes.Interface
 	netServer  MeshDataplane
+	hostIptables       *iptables.IptablesConfigurator
+	hostsideProbeIPSet ipset.IPSet
 }
 
 func (s *meshDataplane) Start(ctx context.Context) {
@@ -198,10 +201,27 @@ func (s *meshDataplane) Start(ctx context.Context) {
 }
 
 func (s *meshDataplane) Stop() {
+	log.Info("CNI ambient server terminating, cleaning up node net rules")
+
+	log.Debug("removing host iptables rules")
+	s.hostIptables.DeleteHostRules()
+
+	log.Debug("destroying host ipset")
+	s.hostsideProbeIPSet.Flush()
+	if err := s.hostsideProbeIPSet.DestroySet(); err != nil {
+		log.Warnf("could not destroy host ipset on shutdown")
+	}
+
 	s.netServer.Stop()
 }
 
 func (s *meshDataplane) ConstructInitialSnapshot(ambientPods []*corev1.Pod) error {
+
+	if err := s.syncHostIPSets(ambientPods); err != nil {
+		log.Errorf("failed to sync host IPset: %v", err)
+		return err
+	}
+
 	return s.netServer.ConstructInitialSnapshot(ambientPods)
 }
 
@@ -238,7 +258,7 @@ func (s *meshDataplane) AddPodToMesh(ctx context.Context, pod *corev1.Pod, podIP
 	// To fix this we probably need to impl CmdDel + manage our own podUID/IP mapping.
 	if retErr == nil {
 		// Handle node healthcheck probe rewrites
-		_, err = s.netServer.AddPodToHostNSIpset(pod, podIPs)
+		_, err = s.addPodToHostNSIpset(pod, podIPs)
 		if err != nil {
 			log.Errorf("failed to add pod to ipset: %s/%s %v", pod.Namespace, pod.Name, err)
 			return err
@@ -252,6 +272,14 @@ func (s *meshDataplane) AddPodToMesh(ctx context.Context, pod *corev1.Pod, podIP
 
 func (s *meshDataplane) RemovePodFromMesh(ctx context.Context, pod *corev1.Pod, isDelete bool) error {
 	log := log.WithLabels("ns", pod.Namespace, "name", pod.Name)
+
+	// Remove the hostside ipset entry first, and unconditionally - if later failures happen, we never
+	// want to leave stale entries
+	if err := removePodFromHostNSIpset(pod, &s.hostsideProbeIPSet); err != nil {
+		log.Errorf("failed to remove pod %s from host ipset, error was: %v", pod.Name, err)
+		return err
+	}
+
 	err := s.netServer.RemovePodFromMesh(ctx, pod, isDelete)
 	if err != nil {
 		log.Errorf("failed to remove pod from mesh: %v", err)
@@ -263,4 +291,99 @@ func (s *meshDataplane) RemovePodFromMesh(ctx context.Context, pod *corev1.Pod, 
 		log.Errorf("failed to annotate pod unenrollment: %v", err)
 	}
 	return err
+}
+
+// syncHostIPSets is called after the host node ipset has been created (or found + flushed)
+// during initial snapshot creation, it will insert every snapshotted pod's IP into the set.
+//
+// The set does not allow dupes (obviously, that would be undefined) - but in the real world due to misconfigured
+// IPAM or other things, we may see two pods with the same IP on the same node - we will skip the dupes,
+// which is all we can do - these pods will fail healthcheck until the IPAM issue is resolved (which seems reasonable)
+func (s *meshDataplane) syncHostIPSets(ambientPods []*corev1.Pod) error {
+	var addedIPSnapshot []netip.Addr
+
+	for _, pod := range ambientPods {
+		podIPs := util.GetPodIPsIfPresent(pod)
+		if len(podIPs) == 0 {
+			log.Warnf("pod %s does not appear to have any assigned IPs, not syncing with ipset", pod.Name)
+		} else {
+			addedIps, err := s.addPodToHostNSIpset(pod, podIPs)
+			if err != nil {
+				log.Errorf("pod %s has IP collision, pod will be skipped and will fail healthchecks", pod.Name, podIPs)
+			}
+			addedIPSnapshot = append(addedIPSnapshot, addedIps...)
+		}
+
+	}
+	return pruneHostIPset(sets.New(addedIPSnapshot...), &s.hostsideProbeIPSet)
+}
+
+// addPodToHostNSIpset:
+// 1. get pod manifest
+// 2. Get all pod ips (might be several, v6/v4)
+// 3. update ipsets accordingly
+// 4. return the ones we added successfully, and errors for any we couldn't (dupes)
+//
+// Dupe IPs should be considered an IPAM error and should never happen.
+func (s *meshDataplane) addPodToHostNSIpset(pod *corev1.Pod, podIPs []netip.Addr) ([]netip.Addr, error) {
+	// Add the pod UID as an ipset entry comment, so we can (more) easily find and delete
+	// all relevant entries for a pod later.
+	podUID := string(pod.ObjectMeta.UID)
+	ipProto := uint8(unix.IPPROTO_TCP)
+
+	var ipsetAddrErrs []error
+	var addedIps []netip.Addr
+
+	// For each pod IP
+	for _, pip := range podIPs {
+		// Add to host ipset
+		log.Debugf("adding pod %s probe to ipset %s with ip %s", pod.Name, s.hostsideProbeIPSet.Prefix, pip)
+		// Add IP/port combo to set. Note that we set Replace to false here - we _did_ previously
+		// set it to true, but in theory that could mask weird scenarios where K8S triggers events out of order ->
+		// an add(sameIPreused) then delete(originalIP).
+		// Which will result in the new pod starting to fail healthchecks.
+		//
+		// Since we purge on restart of CNI, and remove pod IPs from the set on every pod removal/deletion,
+		// we _shouldn't_ get any overwrite/overlap, unless something is wrong and we are asked to add
+		// a pod by an IP we already have in the set (which will give an error, which we want).
+		if err := s.hostsideProbeIPSet.AddIP(pip, ipProto, podUID, false); err != nil {
+			ipsetAddrErrs = append(ipsetAddrErrs, err)
+			log.Errorf("failed adding pod %s to ipset %s with ip %s, error was %s",
+				pod.Name, &s.hostsideProbeIPSet.Prefix, pip, err)
+		} else {
+			addedIps = append(addedIps, pip)
+		}
+	}
+
+	return addedIps, errors.Join(ipsetAddrErrs...)
+}
+
+func removePodFromHostNSIpset(pod *corev1.Pod, hostsideProbeSet *ipset.IPSet) error {
+	podIPs := util.GetPodIPsIfPresent(pod)
+	for _, pip := range podIPs {
+		if err := hostsideProbeSet.ClearEntriesWithIP(pip); err != nil {
+			return err
+		}
+		log.Debugf("removed pod name %s with UID %s from host ipset %s by ip %s", pod.Name, pod.UID, hostsideProbeSet.Prefix, pip)
+	}
+
+	return nil
+}
+
+func pruneHostIPset(expected sets.Set[netip.Addr], hostsideProbeSet *ipset.IPSet) error {
+	actualIPSetContents, err := hostsideProbeSet.ListEntriesByIP()
+	if err != nil {
+		log.Warnf("unable to list IPSet: %v", err)
+		return err
+	}
+	actual := sets.New[netip.Addr](actualIPSetContents...)
+	stales := actual.DifferenceInPlace(expected)
+
+	for staleIP := range stales {
+		if err := hostsideProbeSet.ClearEntriesWithIP(staleIP); err != nil {
+			return err
+		}
+		log.Debugf("removed stale ip %s from host ipset %s", staleIP, hostsideProbeSet.Prefix)
+	}
+	return nil
 }
